TestCapacityScheduler: node management tests moved to a separate file
